refactor: update user and balance handling, add two-factor auth features

+++ src/Auth/Infrastructure/Http/Controllers/LogoutController.php
@@ -19,7 +19,7 @@ readonly class LogoutController
     {
         $useCase = new LogoutUseCase(
             new LogoutDTO(
-                Auth::id()
+                (int) Auth::id()
             ),
             $this->userRepository
         );
